Updated Client & Server
Added struct to pass IP address and thread number to the run client function. Declared a global file variable that all three threads will use. Created an init message so the client will send the server it's writer number upon creation. At the moment, the server is not placing anything in shared memory, only printing the output of the data it has received.