Lots of changes
Created header files, initialized shared memory as global, writers send their names (Writer1, Writer2, etc.), writers are sending data from file to server. Current problems are that with three server threads, the writers names are not all printed, need to make a temp buffer of length 15 in the server thread to take data from readers and store in the shared memory. Once completed the server needs to send back the shared memory buffer to the writer to print.